Fix broken bulkEditTags and refactor logic of the component
