+++ frontend/src/modules/member/components/list/member-list-toolbar.vue
@@ -95,10 +95,9 @@
       </template>
     </el-dropdown>
 
-    <app-member-list-bulk-update-tags
-      v-model="bulkTagsUpdateVisible"
-      :selected-rows="selectedMembers"
-    />
+    <app-tag-popover v-model="bulkTagsUpdateVisible"
+      @reload="fetchMembers({ reload: true })" />
+
   </div>
 </template>
 
@@ -120,7 +119,7 @@ import {
   getEnrichmentMax,
   showEnrichmentLoadingMessage,
 } from '@/modules/member/member-enrichment';
-import AppMemberListBulkUpdateTags from '@/modules/member/components/list/member-list-bulk-update-tags.vue';
+import AppTagPopover from '@/modules/tag/components/tag-popover.vue';
 import AppSvg from '@/shared/svg/svg.vue';
 
 const { currentUser, currentTenant } = mapGetters('auth');
