+++ frontend/src/modules/tag/components/tag-popover.vue
@@ -1,36 +1,18 @@
 <template>
-  <app-dialog
-    v-if="computedVisible"
-    v-model="computedVisible"
-    title="Edit tags"
-    :pre-title="modelValue?.displayName ?? ''"
-  >
+  <app-dialog v-if="computedVisible" v-model="computedVisible" title="Edit tags">
     <template #content>
       <div class="px-6 pb-6">
         <form v-if="modelValue" class="tags-form">
-          <app-tag-autocomplete-input
-            v-model="form"
-            :fetch-fn="fields.tags.fetchFn"
-            :mapper-fn="fields.tags.mapperFn"
-            :create-if-not-found="true"
-            placeholder="Type to search/create tags"
-          />
+          <app-tag-autocomplete-input v-model="bulkEditTagsModel" :fetch-fn="fields.tags.fetchFn"
+            :mapper-fn="fields.tags.mapperFn" :create-if-not-found="true" placeholder="Type to search/create tags" />
         </form>
       </div>
 
-      <div
-        class="bg-gray-50 rounded-b-md flex items-center justify-end py-4 px-6"
-      >
-        <el-button
-          class="btn btn--bordered btn--md mr-3"
-          @click="computedVisible = false"
-        >
+      <div class="bg-gray-50 rounded-b-md flex items-center justify-end py-4 px-6">
+        <el-button class="btn btn--bordered btn--md mr-3" @click="handleCancel">
           Cancel
         </el-button>
-        <el-button
-          class="btn btn--primary btn--md"
-          @click="handleSubmit"
-        >
+        <el-button class="btn btn--primary btn--md" @click="handleSubmit">
           Submit
         </el-button>
       </div>
@@ -42,9 +24,17 @@
 import { MemberModel } from '@/modules/member/member-model';
 import AppDialog from '@/shared/dialog/dialog.vue';
 import AppTagAutocompleteInput from '@/modules/tag/components/tag-autocomplete-input.vue';
+import { FormSchema } from '@/shared/form/form-schema';
 import { mapActions } from 'vuex';
+import { storeToRefs } from 'pinia';
+import { useMemberStore } from '@/modules/member/store/pinia';
+
+const memberStore = useMemberStore();
+const { selectedMembers } = storeToRefs(memberStore);
 
 const { fields } = MemberModel;
+const formSchema = new FormSchema([fields.tags]);
+
 
 export default {
   name: 'AppTagPopover',
@@ -52,17 +42,17 @@ export default {
 
   props: {
     modelValue: {
-      type: Object,
-      default: () => null,
+      type: Boolean,
+      default: () => false,
     },
   },
   emits: ['reload', 'update:modelValue'],
 
   data() {
     return {
-      changed: false,
       loading: false,
-      form: [],
+      bulkEditTagsModel: [],
+      bulkEditTagsInCommon: [],
     };
   },
 
@@ -72,20 +62,19 @@ export default {
     },
     computedVisible: {
       get() {
-        return this.modelValue !== null;
+        return this.modelValue;
       },
       set() {
-        this.$emit('update:modelValue', null);
+        this.$emit('update:modelValue', false);
       },
     },
   },
 
   watch: {
     modelValue: {
-      immediate: true,
-      handler(member) {
-        if (member) {
-          this.form = member.tags;
+      async handler(newValue) {
+        if (newValue) {
+          await this.prepareBulkUpdateTags();
         }
       },
     },
@@ -93,20 +82,51 @@ export default {
 
   methods: {
     ...mapActions({
-      doUpdate: 'member/doUpdate',
+      doBulkUpdateMembersTags:
+        'member/doBulkUpdateMembersTags',
     }),
+
+    prepareBulkUpdateTags() {
+      this.bulkEditTagsModel = selectedMembers.value.reduce(
+        (acc, item, index) => {
+          let { tags } = formSchema.initialValues({
+            tags: item.tags,
+          });
+          if (index > 0) {
+            tags = tags.filter(
+              (tag) => acc.filter((t) => t.id === tag.id).length
+                > 0,
+            );
+          }
+          return tags;
+        },
+        [],
+      );
+      this.bulkEditTagsInCommon = [
+        ...this.bulkEditTagsModel,
+      ];
+    },
+
     async handleSubmit() {
       this.loading = true;
-      await this.doUpdate({
-        id: this.modelValue.id,
-        values: {
-          tags: this.form.map((tag) => tag.id),
-        },
+
+      await this.doBulkUpdateMembersTags({
+        members: [...selectedMembers.value],
+        tagsInCommon: this.bulkEditTagsInCommon,
+        tagsToSave: this.bulkEditTagsModel,
       });
+
       this.loading = false;
       this.computedVisible = false;
-      this.$emit('reload');
+      this.$emit('reload', true);
+      return null;
     },
+
+    handleCancel() {
+      this.bulkEditTagsModel = [];
+      this.bulkEditTagsInCommon = [];
+      this.computedVisible = false;
+    }
   },
 };
 </script>
